requirements.txt
streamlit
numpy
pillow
matplotlib
pandas
streamlit-folium
folium
pystac-client
shapely
pyproj
rasterio
requests

scikit-learn

pyshp
